+++ paddle/phi/kernels/xpu/log_loss_kernel.cc
@@ -29,7 +29,7 @@ void LogLossXPUKernel(const Context& dev_ctx,
   auto* loss = out;
   auto epsilon = static_cast<T>(epsilon_in);
   dev_ctx.template Alloc<T>(loss);
-  int n = predict->numel();
+  int64_t n = predict->numel();
   int r = xpu::log_loss(dev_ctx.x_context(),
                         predict->data<T>(),
                         labels->data<T>(),
